refactors Thumbnail, Gallery, and Selector to separate files imports Thumbnail, Gallery, and Selector to index returns figures in Thumbnail and Gallery
